+++ src/ott/solvers/nn/__init__.py
@@ -1,1 +1,1 @@
-from . import icnn, layers, neuraldual
+from . import conjugate_solver, layers, models, neuraldual
