+++ src/ott/solvers/nn/models.py
@@ -1,5 +1,3 @@
-# Copyright 2022 Google LLC.
-#
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
@@ -11,25 +9,114 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Implementation of :cite:`amos:17` input convex neural networks (ICNN)."""
+"""Base module for potentials."""
 
+import abc
 from typing import Callable, Optional, Sequence, Tuple, Union
 
 import flax.linen as nn
 import jax
 import jax.numpy as jnp
 import optax
+from flax import struct
 from flax.core import frozen_dict
+from flax.training import train_state
 from jax.nn import initializers
 
 from ott.math import matrix_square_root
-from ott.solvers.nn import potential_base
 from ott.solvers.nn.layers import PosDefPotentials, PositiveDense
 
-__all__ = ["ICNN"]
+__all__ = ["NeuralTrainState", "ModelBase", "ICNN", "MLP"]
+
+PotentialValueFn_t = Callable[[jnp.ndarray], jnp.ndarray]
+PotentialGradientFn_t = Callable[[jnp.ndarray], jnp.ndarray]
 
 
-class ICNN(potential_base.PotentialBase):
+class NeuralTrainState(train_state.TrainState):
+  """Adds information about the model's value and gradient to the state."""
+  potential_value_fn: Callable[
+      [frozen_dict.FrozenDict[str, jnp.ndarray], Optional[PotentialValueFn_t]],
+      PotentialValueFn_t] = struct.field(pytree_node=False)
+  potential_gradient_fn: Callable[[frozen_dict.FrozenDict[str, jnp.ndarray]],
+                                  PotentialGradientFn_t] = struct.field(
+                                      pytree_node=False
+                                  )
+
+
+class ModelBase(abc.ABC, nn.Module):
+  """Base class for the neural solver models."""
+
+  @property
+  @abc.abstractmethod
+  def is_potential(self) -> bool:
+    """Indicates if the module defines the potential's value or the gradient.
+
+    Returns:
+      ``True`` if the module defines the potential's value, ``False``
+      if it defines the gradient.
+    """
+
+  def potential_value_fn(
+      self,
+      params: frozen_dict.FrozenDict[str, jnp.ndarray],
+      other_potential_value_fn: Optional[PotentialValueFn_t] = None,
+  ) -> PotentialValueFn_t:
+    r"""Return a function giving the value of the potential.
+
+    Applies the module if ``self.is_potential`` is ``True``, otherwise
+    constructs the value of the potential from the gradient with
+
+    .. math::
+      g(y) = -f(\nabla_y g(y)) + y^T \nabla_y g(y)
+
+    where :math:`\nabla_y g(y)` is detached for the envelope theorem
+    :cite:`danskin:67,bertsekas:71`
+    to give the appropriate first derivatives of this construction.
+
+    Args:
+      params: parameters of the module
+      x: point to evaluate the value at
+      other_potential_value: function giving the value of the other potential.
+        Only needed when ``self.is_potential`` is ``False``.
+
+    Returns:
+      A function that can be evaluated to obtain the potential's value
+    """
+    if self.is_potential:
+      return lambda x: self.apply({"params": params}, x)
+    else:
+      assert other_potential_value_fn is not None
+
+      def value_fn(x):
+        squeeze = x.ndim == 1
+        if squeeze:
+          x = jnp.expand_dims(x, 0)
+        grad_g_x = jax.lax.stop_gradient(self.apply({"params": params}, x))
+        value = -other_potential_value_fn(grad_g_x) + \
+            jax.vmap(jnp.dot)(grad_g_x, x)
+        return value.squeeze(0) if squeeze else value
+
+      return value_fn
+
+  def potential_gradient_fn(
+      self,
+      params: frozen_dict.FrozenDict[str, jnp.ndarray],
+  ) -> PotentialGradientFn_t:
+    """Return a function giving the gradient of the potential.
+
+    Args:
+      params: parameters of the module
+
+    Returns:
+      A function that can be evaluated to obtain the potential's gradient
+    """
+    if self.is_potential:
+      return jax.vmap(jax.grad(self.potential_value_fn(params)))
+    else:
+      return lambda x: self.apply({'params': params}, x)
+
+
+class ICNN(ModelBase):
   """Input convex neural network (ICNN) architecture with initialization.
 
   Implementation of input convex neural networks as introduced in
@@ -191,10 +278,71 @@ class ICNN(potential_base.PotentialBase):
       optimizer: optax.OptState,
       input: Union[int, Tuple[int, ...]],
       params: Optional[frozen_dict.FrozenDict[str, jnp.ndarray]] = None,
-  ) -> potential_base.PotentialTrainState:
+  ) -> NeuralTrainState:
     """Create initial `TrainState`."""
     params = self.init(rng, jnp.ones(input))["params"]
-    return potential_base.PotentialTrainState.create(
+    return NeuralTrainState.create(
+        apply_fn=self.apply,
+        params=params,
+        tx=optimizer,
+        potential_value_fn=self.potential_value_fn,
+        potential_gradient_fn=self.potential_gradient_fn,
+    )
+
+
+class MLP(ModelBase):
+  """A non-convex MLP.
+
+  Args:
+    dim_hidden: sequence specifying size of hidden dimensions. The
+      output dimension of the last layer is automatically set to
+      1 (if `is_potential=True`) or the dimension
+      of the input (if `is_potential=False`).
+    is_potential: Model the potential if `True`, otherwise
+      model the gradient of the potential
+    act_fn: Activation function
+  """
+
+  dim_hidden: Sequence[int]
+  is_potential: bool = True
+  act_fn: Callable[[jnp.ndarray], jnp.ndarray] = nn.leaky_relu
+
+  @nn.compact
+  def __call__(self, x: jnp.ndarray) -> jnp.ndarray:
+    squeeze = x.ndim == 1
+    if squeeze:
+      x = jnp.expand_dims(x, 0)
+    assert x.ndim == 2, x.ndim
+    n_input = x.shape[-1]
+
+    z = x
+    for n_hidden in self.dim_hidden:
+      Wx = nn.Dense(n_hidden, use_bias=True)
+      z = self.act_fn(Wx(z))
+
+    if self.is_potential:
+      Wx = nn.Dense(1, use_bias=True)
+      z = Wx(z).squeeze(-1)
+
+      quad_term = 0.5 * jax.vmap(jnp.dot)(x, x)
+      z += quad_term
+    else:
+      Wx = nn.Dense(n_input, use_bias=True)
+      z = x + Wx(z)
+
+    return z.squeeze(0) if squeeze else z
+
+  def create_train_state(
+      self,
+      rng: jnp.ndarray,
+      optimizer: optax.OptState,
+      input: Union[int, Tuple[int, ...]],
+      params: Optional[frozen_dict.FrozenDict[str, jnp.ndarray]] = None,
+  ) -> NeuralTrainState:
+    """Create initial `TrainState`."""
+    if params is None:
+      params = self.init(rng, jnp.ones(input))["params"]
+    return NeuralTrainState.create(
         apply_fn=self.apply,
         params=params,
         tx=optimizer,
